ci/cd support: export some go routines for use in ci/cd
Useful for integration testing from within golang test files.
